Manually render DIsableSelectionClipping and ForceAliasedSelectionQuality
**NOTE:** See my comment on line 142 about the use of `CloneWithAliasedSelection`. I think this may need to be added elsewhere too, but I don't have the time to look for it right now.

This makes it so that CodeLab is responsible for implementing these two flags. Otherwise there are discrepancies that pop up when rendering classic effects, as seen at https://forums.getpaint.net/topic/123490-paintnet-5011-alpha-build-8683/?do=findComment&comment=613825

CodeLab will now _always_ specify `DisableSelectionClipping` in its `OnInitializeRenderInfo`, and will _never_ specify `ForceAliasedSelectionQuality`. That gives it full access to the output buffer, and 100% fidelity for the selection.

For a script that wants to use `DisableSelectionClipping`, rendering is very easy -- just render directly to the output buffer. When that flags is not desired, rendering is more involved and also hinges on the value of `ForceAliasedSelectionQuality`.

This also opens things up so that CodeLab can push these settings down into the script, and the user can toggle them without having to reboot CodeLab.